+++ src/screens/Signup/StepCaptcha/CaptchaWebView.web.tsx
