+++ src/screens/Signup/StepCaptcha/index.tsx
@@ -6,9 +6,9 @@ import {nanoid} from 'nanoid/non-secure'
 
 import {createFullHandle} from '#/lib/strings/handles'
 import {isWeb} from '#/platform/detection'
-import {CaptchaWebView} from '#/view/com/auth/create/CaptchaWebView'
 import {ScreenTransition} from '#/screens/Login/ScreenTransition'
 import {useSignupContext, useSubmitSignup} from '#/screens/Signup/state'
+import {CaptchaWebView} from '#/screens/Signup/StepCaptcha/CaptchaWebView'
 import {atoms as a, useTheme} from '#/alf'
 import {FormError} from '#/components/forms/FormError'
