+++ src/screens/Signup/StepInfo/Policies.tsx
