+++ src/screens/Signup/StepInfo/index.tsx
@@ -4,9 +4,9 @@ import {msg, Trans} from '@lingui/macro'
 import {useLingui} from '@lingui/react'
 
 import {logger} from '#/logger'
-import {Policies} from 'view/com/auth/create/Policies'
 import {ScreenTransition} from '#/screens/Login/ScreenTransition'
 import {is13, is18, useSignupContext} from '#/screens/Signup/state'
+import {Policies} from '#/screens/Signup/StepInfo/Policies'
 import {atoms as a} from '#/alf'
 import * as DateField from '#/components/forms/DateField'
 import {FormError} from '#/components/forms/FormError'
